src/leo_skills/content_creation/realestate_news_publisher_skill/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.10

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# AI / Content Generation
zhipuai>=2.0.0
openai>=1.0.0

# WeChat Integration
wechatpy>=1.8.18

# Database
sqlalchemy>=2.0.0

# Async support
aiohttp>=3.9.0
aiofiles>=23.0.0

# Text processing
jieba>=0.42.1
scikit-learn>=1.3.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.66.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
